Reworked files into utils module, cleaned up file headers
