+++ crucible/utils/__init__.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Crucible utility package.
+
+Submodules
+----------
+io          : File I/O, hashing, time, and image helpers
+deprecation : API lifecycle decorators (_deprecated, _removed)
+"""
+
+from .io import run_shell, checkhash, get_tz_isoformat, check_small_files, data2thumbnail
+from .deprecation import _deprecated, _removed
+
+__all__ = [
+    # io
+    'run_shell',
+    'checkhash',
+    'get_tz_isoformat',
+    'check_small_files',
+    'data2thumbnail',
+    # deprecation
+    '_deprecated',
+    '_removed',
+]
